+++ gazebo/physics/WorldPrivate.hh
@@ -243,6 +243,10 @@ namespace gazebo
       /// \brief Buffer of prev states
       public: WorldState prevStates[2];
 
+      /// \brief Previous unfiltered state. Used for determining insertions
+      /// and deletions
+      public: WorldState prevUnfilteredState;
+
       /// \brief Int used to toggle between prevStates
       public: int stateToggle;
 
@@ -323,6 +327,9 @@ namespace gazebo
       /// \brief True if sensors have been initialized. This should be set
       /// by the SensorManager.
       public: std::atomic_bool sensorsInitialized;
+
+      /// \brief Simulation time of the last log state captured.
+      public: gazebo::common::Time logLastStateTime;
     };
   }
 }